refactor: rename graph service file
