+++ tests/board-utils.test.ts
@@ -1,22 +1,26 @@
 import { forEachSelection } from '../src/board/board';
 
 describe('forEachSelection', () => {
-  test('invokes callback for each item', async () => {
-    const items = [{ a: 1 }, { b: 2 }];
-    const board = { getSelection: jest.fn().mockResolvedValue(items) };
-    const cb = jest.fn();
-    await forEachSelection(cb, board);
-    expect(cb).toHaveBeenCalledTimes(2);
-    expect(cb).toHaveBeenCalledWith(items[0]);
-    expect(cb).toHaveBeenCalledWith(items[1]);
+  describe('callback invocation', () => {
+    test('invokes callback once per item', async () => {
+      const items = [{ a: 1 }, { b: 2 }];
+      const board = { getSelection: jest.fn().mockResolvedValue(items) };
+      const cb = jest.fn();
+      await forEachSelection(cb, board);
+      expect(cb).toHaveBeenCalledTimes(items.length);
+      expect(cb).toHaveBeenNthCalledWith(1, items[0]);
+      expect(cb).toHaveBeenNthCalledWith(2, items[1]);
+    });
   });
 
-  test('propagates callback errors', async () => {
-    const board = { getSelection: jest.fn().mockResolvedValue([{}]) };
-    await expect(
-      forEachSelection(() => {
-        throw new Error('fail');
-      }, board),
-    ).rejects.toThrow('fail');
+  describe('error propagation', () => {
+    test('rejects when callback throws', async () => {
+      const board = { getSelection: jest.fn().mockResolvedValue([{}]) };
+      await expect(
+        forEachSelection(() => {
+          throw new Error('fail');
+        }, board),
+      ).rejects.toThrow('fail');
+    });
   });
 });
